:hammer: Replace QA workflow with temporary deployment command.
