+++ CHANGELOG.md
@@ -6,4 +6,4 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Added a local `bin/qa` workflow for isolated per-PR Azure Container Apps QA environments, including readiness checks, persisted cleanup state, and non-production resource guardrails.
+- Added `bin/deploy temporary` for disposable public Azure Container Apps environments from pull requests, branches, or immutable git hashes, including readiness checks, persisted cleanup state, and non-production resource guardrails.
