+++ Dockerfile
@@ -38,15 +38,15 @@ RUN npm clean-install
 
 COPY web ./
 
-ARG VITE_QA_HOST_SUFFIX
-ARG VITE_QA_AUTH0_DOMAIN
-ARG VITE_QA_AUTH0_AUDIENCE
-ARG VITE_QA_AUTH0_CLIENT_ID
+ARG VITE_TEMPORARY_HOST_SUFFIX
+ARG VITE_TEMPORARY_AUTH0_DOMAIN
+ARG VITE_TEMPORARY_AUTH0_AUDIENCE
+ARG VITE_TEMPORARY_AUTH0_CLIENT_ID
 ENV NODE_ENV=production
-ENV VITE_QA_HOST_SUFFIX=${VITE_QA_HOST_SUFFIX}
-ENV VITE_QA_AUTH0_DOMAIN=${VITE_QA_AUTH0_DOMAIN}
-ENV VITE_QA_AUTH0_AUDIENCE=${VITE_QA_AUTH0_AUDIENCE}
-ENV VITE_QA_AUTH0_CLIENT_ID=${VITE_QA_AUTH0_CLIENT_ID}
+ENV VITE_TEMPORARY_HOST_SUFFIX=${VITE_TEMPORARY_HOST_SUFFIX}
+ENV VITE_TEMPORARY_AUTH0_DOMAIN=${VITE_TEMPORARY_AUTH0_DOMAIN}
+ENV VITE_TEMPORARY_AUTH0_AUDIENCE=${VITE_TEMPORARY_AUTH0_AUDIENCE}
+ENV VITE_TEMPORARY_AUTH0_CLIENT_ID=${VITE_TEMPORARY_AUTH0_CLIENT_ID}
 RUN npm run build
 
 # Stage 3 - production setup
