+++ api/src/app.ts
@@ -4,7 +4,7 @@ import path from "path"
 import helmet from "helmet"
 import formData from "express-form-data"
 
-import { AUTH0_DOMAIN, FRONTEND_URL, GIT_COMMIT_HASH, QA_ENVIRONMENT, RELEASE_TAG } from "@/config"
+import { AUTH0_DOMAIN, FRONTEND_URL } from "@/config"
 import { betterFormDataBodyParserMiddleware, requestLoggerMiddleware } from "@/middlewares"
 import router from "@/router"
 import enhancedQsDecoder from "@/utils/enhanced-qs-decoder"
@@ -46,11 +46,6 @@ app.use(
 
 app.use(requestLoggerMiddleware)
 
-if (QA_ENVIRONMENT) {
-  app.get("/qa-status", (_request: Request, response: Response) => {
-    response.json({ status: "ok", releaseTag: RELEASE_TAG, gitCommitHash: GIT_COMMIT_HASH })
-  })
-}
 app.use(router)
 
 // serves the static files generated by the front-end
