+++ api/src/config.ts
@@ -48,7 +48,6 @@ export const RUN_SCHEDULER = process.env.RUN_SCHEDULER || "false"
 export const AUTH0_DOMAIN = stripTrailingSlash(process.env.VITE_AUTH0_DOMAIN || "")
 export const AUTH0_AUDIENCE = process.env.VITE_AUTH0_AUDIENCE
 export const AUTH0_REDIRECT = process.env.VITE_AUTH0_REDIRECT || process.env.FRONTEND_URL || ""
-export const QA_ENVIRONMENT = process.env.QA_ENVIRONMENT === "true"
 // ====================
 // Database & Cache Configuration
 // ====================
