+++ bin/deploy
@@ -1,0 +1,12 @@
+#!/usr/bin/env ruby
+
+require_relative "lib/deploy/all"
+
+if $PROGRAM_NAME == __FILE__
+  begin
+    TraditionalKnowledgeTemporaryDeployment.run(ARGV)
+  rescue ArgumentError => error
+    warn error.message
+    exit 1
+  end
+end
