+++ bin/lib/deploy/all.rb
@@ -1,0 +1,14 @@
+require_relative "temporary/constants"
+require_relative "temporary/support/errors"
+require_relative "temporary/support/state_store"
+require_relative "temporary/support/source"
+require_relative "temporary/support/config"
+require_relative "temporary/support/github"
+require_relative "temporary/support/auth0"
+require_relative "temporary/azure"
+require_relative "temporary/azure/access"
+require_relative "temporary/azure/resources"
+require_relative "temporary/azure/deployment_body"
+require_relative "temporary/support/worktree"
+require_relative "temporary/application"
+require_relative "temporary/command"
