--- bin/qa
@@ -1,5 +1,0 @@
-#!/usr/bin/env ruby
-
-require_relative "lib/qa"
-
-TraditionalKnowledgeQa.run(ARGV) if $PROGRAM_NAME == __FILE__
